requirements.txt
assimp_py==1.0.6
numpy==1.22.3
perlin_noise==1.12
pygame==2.0.1
PyGLM==2.5.7
PyOpenGL==3.1.6
